refactor(wds)!: rehome the session vocabulary, units and results contract
Phase 4 unweave of the format-blind extraction, mirroring the uds
pattern. The session vocabulary — warnings, reported quantities,
balances, snapshots, and the WritableSimulation contract — moves from
io into simulation::contract, flipping a re-export that ran backwards;
unit conversion moves to model::units, being model semantics; and the
§13 analysis reads results through a ResultsSource contract
(analysis::source: neutral metadata, per-period reads, ranged and
streaming scans) instead of file paths — the .out byte layout stays
dialect-side as the OutLayout extension, and OutFileSource implements
the contract over the carve-out. Dialect sources reference
crate::engine_api now, so the coming lift moves them without another
churn. An unknown block id is rejected before any read, and the test
proves it with a source that panics on touch.

BREAKING CHANGE: produce_report_block and the analysis entry points
take a ResultsSource instead of a path; build OutFileSource::open for
the file-backed case.
